feat: /chat/[chatRoomId] 경로 추가에 따른 채팅 구조 변경
